Added commitment parameters to GetBalance, GetAccountInfo and GetBlock requests. Split some of the unit tests into logical partial classes.
