+++ test/Solnet.Rpc.Test/Resources/Http/Accounts/GetAccountInfoConfirmedRequest.json
@@ -1,0 +1,1 @@
+{"method":"getAccountInfo","params":["9we6kjtbcZ2vy3GSLLsZTEhbAqXPTRvEyoxa8wxSqKp5",{"encoding":"base64","commitment":"confirmed"}],"jsonrpc":"2.0","id":0}
