+++ test/Solnet.Rpc.Test/Resources/Http/Blocks/GetBlockConfirmedRequest.json
@@ -1,0 +1,1 @@
+{"method":"getBlock","params":[79662905,{"encoding":"json","commitment":"confirmed"}],"jsonrpc":"2.0","id":0}
