+++ test/Solnet.Rpc.Test/Resources/Http/Blocks/GetBlockRequest.json
@@ -1,1 +1,1 @@
-{"method":"getBlock","params":[79662905,{"encoding":"json","transactionDetails":"full"}],"jsonrpc":"2.0","id":0}
+{"method":"getBlock","params":[79662905,{"encoding":"json"}],"jsonrpc":"2.0","id":0}
