+++ test/Solnet.Rpc.Test/Resources/Http/GetBalanceConfirmedRequest.json
@@ -1,0 +1,1 @@
+{"method":"getBalance","params":["hoakwpFB8UoLnPpLC56gsjpY7XbVwaCuRQRMQzN5TVh",{"commitment":"confirmed"}],"jsonrpc":"2.0","id":0}
